OBGM-47 Refactor CategoryExcelImporter - rename CategoryDataService to CategoryService - create CategoryDataImportService for validation and import logic
