+++ grails-app/services/org/pih/warehouse/data/CategoryService.groovy
@@ -1,0 +1,28 @@
+/**
+ * Copyright (c) 2012 Partners In Health.  All rights reserved.
+ * The use and distribution terms for this software are covered by the
+ * Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+ * which can be found in the file epl-v10.html at the root of this distribution.
+ * By using this software in any fashion, you are agreeing to be bound by
+ * the terms of this license.
+ * You must not remove this notice, or any other, from this software.
+ **/
+package org.pih.warehouse.data
+
+import grails.gorm.transactions.Transactional
+import org.pih.warehouse.product.Category
+
+@Transactional
+class CategoryService {
+    Category createOrUpdateCategory(Map params) {
+        Category category = Category.findByIdOrName(params.id, params.name)
+        if (!category) {
+            category = new Category()
+        }
+        category.name = params.name
+        if (params.parentCategoryId && category?.parentCategory?.id != params.parentCategoryId) {
+            category.parentCategory = Category.get(params.parentCategoryId)
+        }
+        return category
+    }
+}
